+++ src/UeWebViewer/viewer/ExportDetails.tsx
@@ -48,7 +48,7 @@ function makeTree(asset: Asset): Node[] {
   };
 
   const convertedTable = asset.exports.map((exportObject, index) => {
-    const fullName = asset.makeFullName(exportObject.ClassIndex);
+    const fullName = asset.makeFullNameByIndex(exportObject.ClassIndex);
     return new Node(exportObject, index + 1, fullName);
   });
