+++ src/unreal-engine/__tests__/deserialization/Basic.ts
@@ -23,11 +23,11 @@ describe("TestActorUE4-0-2", () => {
   });
 
   test("makeFullName", () => {
-    expect(asset.makeFullName(0).toString()).toBe("None");
-    expect(asset.makeFullName(55).toString()).toBe(
+    expect(asset.makeFullNameByIndex(0).toString()).toBe("None");
+    expect(asset.makeFullNameByIndex(55).toString()).toBe(
       "TestActorUE4-0-2.TestActorUE4-0-2:CustomFunction.K2Node_CallFunction_7631",
     );
-    expect(asset.makeFullName(-36).toString()).toBe("/Script/CoreUObject.Object:LinearColor");
+    expect(asset.makeFullNameByIndex(-36).toString()).toBe("/Script/CoreUObject.Object:LinearColor");
   });
 });
 
@@ -53,10 +53,10 @@ describe("TestActorUE5-4-4", () => {
   });
 
   test("makeFullName", () => {
-    expect(asset.makeFullName(0).toString()).toBe("None");
-    expect(asset.makeFullName(14).toString()).toBe(
+    expect(asset.makeFullNameByIndex(0).toString()).toBe("None");
+    expect(asset.makeFullNameByIndex(14).toString()).toBe(
       "TestActorUE5-4-4.TestActorUE5-4-4:CustomGraph.K2Node_CallFunction_0",
     );
-    expect(asset.makeFullName(-21).toString()).toBe("/Script/Engine.KismetSystemLibrary:PrintString");
+    expect(asset.makeFullNameByIndex(-21).toString()).toBe("/Script/Engine.KismetSystemLibrary:PrintString");
   });
 });
