+++ src/unreal-engine/modules/CoreUObject/objects/Class.ts
