+++ src/unreal-engine/modules/CoreUObject/objects/Package.ts
