+++ src/unreal-engine/modules/CoreUObject/structs/Box.ts
@@ -1,4 +1,4 @@
-import type { AssetReader } from "../../AssetReader";
+import type { AssetReader } from "../../../AssetReader";
 import { FVector3 } from "./Vector3";
 
 export class FBox {
