+++ src/unreal-engine/modules/CoreUObject/structs/Matrix44.ts
@@ -1,4 +1,4 @@
-import type { AssetReader } from "../../AssetReader";
+import type { AssetReader } from "../../../AssetReader";
 
 export type ArrayType = [
   number, // M[0][0]
