+++ src/unreal-engine/modules/CoreUObject/structs/Rotator.ts
@@ -1,4 +1,4 @@
-import type { AssetReader } from "../../AssetReader";
+import type { AssetReader } from "../../../AssetReader";
 
 /**
  * struct FRotator {
