+++ src/unreal-engine/modules/CoreUObject/structs/TwoVectors.ts
@@ -1,4 +1,4 @@
-import type { AssetReader } from "../../AssetReader";
+import type { AssetReader } from "../../../AssetReader";
 import { FVector3 } from "./Vector3";
 
 /**
