+++ src/unreal-engine/modules/CoreUObject/structs/Vector4.ts
@@ -1,4 +1,4 @@
-import type { AssetReader } from "../../AssetReader";
+import type { AssetReader } from "../../../AssetReader";
 
 /**
  * Four-dimensional vector.
