+++ src/unreal-engine/properties/PropertyTag.ts
@@ -1,5 +1,5 @@
 import { FName, NAME_None } from "../structs/Name";
-import { EGuidFormats, FGuid } from "../objects/CoreUObject/Guid";
+import { EGuidFormats, FGuid } from "../modules/CoreUObject/structs/Guid";
 import type { AssetReader } from "../AssetReader";
 import { EUnrealEngineObjectUE4Version, EUnrealEngineObjectUE5Version } from "../versioning/ue-versions";
 import invariant from "tiny-invariant";
