+++ src/unreal-engine/properties/properties-serialization.ts
@@ -4,7 +4,7 @@ import type { AssetReader } from "../AssetReader";
 import { FPropertyTag } from "./PropertyTag";
 import { EPropertyTagExtension, EPropertyType } from "./enums";
 import { EUnrealEngineObjectUE5Version } from "../versioning/ue-versions";
-import type { ObjectResolver } from "../objects/CoreUObject/Object";
+import type { ObjectResolver } from "../modules/CoreUObject/objects/Object";
 import { getPropertySerializerFromTag, UnknownPropertyType } from "./serializers";
 
 export function readTaggedProperties(reader: AssetReader, isUClass: boolean, resolver: ObjectResolver) {
