+++ src/unreal-engine/structs/PackageFileSummary.ts
@@ -1,6 +1,6 @@
 import type { AssetReader } from "../AssetReader";
 import { EPackageFlags } from "../enums";
-import { FGuid, GUID_None } from "../objects/CoreUObject/Guid";
+import { FGuid, GUID_None } from "../modules/CoreUObject/structs/Guid";
 import { FEngineVersion } from "./EngineVersion";
 import { ECustomVersionSerializationFormat, FCustomVersionContainer } from "../serialization/CustomVersion";
 import { EUnrealEngineObjectUE4Version, EUnrealEngineObjectUE5Version } from "../versioning/ue-versions";
