+++ src/unreal-engine/versioning/registry.ts
@@ -1,4 +1,4 @@
-import type { FGuid } from "../objects/CoreUObject/Guid";
+import type { FGuid } from "../modules/CoreUObject/structs/Guid";
 import invariant from "tiny-invariant";
 
 /**
